requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
